first version of sale saving

+++ domain/model/src/main/java/com/softlond/model/product/Product.java
@@ -11,7 +11,7 @@ import lombok.*;
 public class Product {
     private String id;
     private String name;
-    private Double price;
-    private Integer stock;
+    private long price;
+    private int stock;
     private String categoryId;
 }
